Imitation_Learning/requirements.txt
cycler==0.10.0
dataclasses==0.6
future==0.18.2
kiwisolver==1.3.1
matplotlib==3.3.3
numpy==1.19.4
opencv-python==4.4.0.46
Pillow==8.0.1
pybullet==3.0.6
pycodestyle==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
scipy==1.5.4
six==1.15.0
toml==0.10.2
torch==1.7.0
torchvision==0.8.1
tqdm==4.53.0
transformations==2020.1.1
typing-extensions==3.7.4.3
